Add files via upload
ISSUE #1
не все готово, из MainGame перенес потихоньку, но еще есть что делать, у меингейм больно много зависимостей от Drawing все еще
